Initial new structure + moved grid initialization into its own class.
