+++ CMakeLists.txt
@@ -9,7 +9,10 @@ link_directories(./lib)
 
 add_executable(active-microemulsion
         src/main.cpp
-        src/Timing.cpp src/Timing.h
-        src/Logger.cpp src/Logger.h
-        src/Grid.cpp src/Grid.h src/PgmWriter.cpp src/PgmWriter.h src/Microemulsion.cpp src/Microemulsion.h src/CellData.h src/ChainConfig.cpp src/ChainConfig.h src/EventSchedule.cpp src/EventSchedule.h)
+        src/Timing/Timing.cpp src/Timing/Timing.h
+        src/Logger/Logger.cpp src/Logger/Logger.h
+        src/Grid/Grid.cpp src/Grid/Grid.h
+        src/Grid/GridInitializer.cpp src/Grid/GridInitializer.h
+        src/Visualization/PgmWriter.cpp src/Visualization/PgmWriter.h
+        src/Microemulsion/Microemulsion.cpp src/Microemulsion/Microemulsion.h src/Cell/CellData.h src/Chain/ChainConfig.cpp src/Chain/ChainConfig.h src/EventSchedule/EventSchedule.cpp src/EventSchedule/EventSchedule.h)
 target_link_libraries(active-microemulsion boost_program_options boost_system boost_filesystem m)
